+++ src/main/java/com/example/moviereserve/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.example.moviereserve.controller;
 
-import com.example.moviereserve.dto.BusinessSignUpRequestDto;
-import com.example.moviereserve.dto.SignInRequestDto;
-import com.example.moviereserve.dto.UserSignUpRequestDto;
+import com.example.moviereserve.dto.sign.BusinessSignUpRequestDto;
+import com.example.moviereserve.dto.sign.SignInRequestDto;
+import com.example.moviereserve.dto.sign.UserSignUpRequestDto;
 import com.example.moviereserve.entity.User;
 import com.example.moviereserve.exception.LoginFailureException;
 import com.example.moviereserve.repository.UserRepository;
