+++ src/main/java/com/example/moviereserve/service/UserSignInService.java
@@ -1,8 +1,8 @@
 package com.example.moviereserve.service;
 
 import com.example.moviereserve.config.jwt.TokenProvider;
-import com.example.moviereserve.dto.SignInRequestDto;
-import com.example.moviereserve.dto.UserSignInResponseDto;
+import com.example.moviereserve.dto.sign.SignInRequestDto;
+import com.example.moviereserve.dto.sign.UserSignInResponseDto;
 import com.example.moviereserve.entity.User;
 import com.example.moviereserve.exception.LoginFailureException;
 import lombok.RequiredArgsConstructor;
